updated for new interfaces

+++ hobbes/hobbes_db.h
@@ -53,7 +53,7 @@ int hdb_create_enclave(hdb_db_t       db,
 		       char         * name, 
 		       int            mgmt_dev_id, 
 		       enclave_type_t type, 
-		       uint64_t            parent);
+		       int            parent);
 
 
 int hdb_update_enclave(hdb_db_t                db,
